Move the macros to separate module
